# Create ExpressionCache to unify constant lookups