--- a/Jint/Runtime/Interpreter/Expressions/JintNewExpression.cs
+++ b/Jint/Runtime/Interpreter/Expressions/JintNewExpression.cs
@@ -1,42 +1,27 @@
-using Jint.Native;
-
 namespace Jint.Runtime.Interpreter.Expressions;
 
 internal sealed class JintNewExpression : JintExpression
 {
+    private readonly ExpressionCache _arguments = new();
     private JintExpression _calleeExpression = null!;
-    private JintExpression[] _jintArguments = Array.Empty<JintExpression>();
-    private bool _hasSpreads;
     private bool _initialized;
 
     public JintNewExpression(NewExpression expression) : base(expression)
     {
     }
 
-    private void Initialize()
+    private void Initialize(EvaluationContext context)
     {
         var expression = (NewExpression) _expression;
+        _arguments.Initialize(context, expression.Arguments.AsSpan());
         _calleeExpression = Build(expression.Callee);
-
-        if (expression.Arguments.Count <= 0)
-        {
-            return;
-        }
-
-        _jintArguments = new JintExpression[expression.Arguments.Count];
-        for (var i = 0; i < _jintArguments.Length; i++)
-        {
-            var argument = expression.Arguments[i];
-            _jintArguments[i] = Build(argument);
-            _hasSpreads |= argument.Type == NodeType.SpreadElement;
-        }
     }
 
     protected override object EvaluateInternal(EvaluationContext context)
     {
         if (!_initialized)
         {
-            Initialize();
+            Initialize(context);
             _initialized = true;
         }
 
@@ -45,20 +30,7 @@ internal sealed class JintNewExpression : JintExpression
         // todo: optimize by defining a common abstract class or interface
         var jsValue = _calleeExpression.GetValue(context);
 
-        JsValue[] arguments;
-        if (_jintArguments.Length == 0)
-        {
-            arguments = Array.Empty<JsValue>();
-        }
-        else if (_hasSpreads)
-        {
-            arguments = BuildArgumentsWithSpreads(context, _jintArguments);
-        }
-        else
-        {
-            arguments = engine._jsValueArrayPool.RentArray(_jintArguments.Length);
-            BuildArguments(context, _jintArguments, arguments);
-        }
+        var arguments = _arguments.ArgumentListEvaluation(context, out var rented);
 
         // Reset the location to the "new" keyword so that if an Error object is
         // constructed below, the stack trace will capture the correct location.
@@ -66,13 +38,16 @@ internal sealed class JintNewExpression : JintExpression
 
         if (!jsValue.IsConstructor)
         {
-            ExceptionHelper.ThrowTypeError(engine.Realm, _calleeExpression.SourceText + " is not a constructor");
+            ExceptionHelper.ThrowTypeError(engine.Realm, $"{_calleeExpression.SourceText} is not a constructor");
         }
 
         // construct the new instance using the Function's constructor method
         var instance = engine.Construct(jsValue, arguments, jsValue, _calleeExpression);
 
-        engine._jsValueArrayPool.ReturnArray(arguments);
+        if (rented)
+        {
+            engine._jsValueArrayPool.ReturnArray(arguments);
+        }
 
         return instance;
     }