… Jint/Runtime/Interpreter/Expressions/JintObjectExpression.cs