Puts common logic and duplicated code to one place.