@lahma 2025-01-24
No major differences in performance.

## Jint.Benchmark.DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Run|dromaeo-3d-cube|15.778 ms|0.2194 ms|6303.53 KB|
| **New** |	|	| **15.480 ms (-2%)** | **0.1574 ms** | **6305.56 KB (0%)** |
| Old |Run|dromaeo-core-eval|3.090 ms|0.0325 ms|327.5 KB|
| **New** |	|	| **3.002 ms (-3%)** | **0.0252 ms** | **327.42 KB (0%)** |
| Old |Run|dromaeo-object-array|31.589 ms|0.4030 ms|96259.27 KB|
| **New** |	|	| **32.026 ms (+1%)** | **0.2861 ms** | **96259.08 KB (0%)** |
| Old |Run|droma(...)egexp [21]|125.476 ms|2.5058 ms|150032.96 KB|
| **New** |	|	| **121.780 ms (-3%)** | **2.3399 ms** | **150597.65 KB (0%)** |
| Old |Run|droma(...)tring [21]|239.919 ms|5.4312 ms|1315983.34 KB|
| **New** |	|	| **233.933 ms (-2%)** | **4.9587 ms** | **1315737.61 KB (0%)** |
| Old |Run|droma(...)ase64 [21]|35.003 ms|0.5498 ms|2367.54 KB|
| **New** |	|	| **35.679 ms (+2%)** | **0.6749 ms** | **2367.42 KB (0%)** |
| Old |Run|dromaeo-3d-cube|15.177 ms|0.1215 ms|5965.28 KB|
| **New** |	|	| **15.202 ms (0%)** | **0.1546 ms** | **5967.31 KB (0%)** |
| Old |Run|dromaeo-core-eval|3.032 ms|0.0233 ms|312.43 KB|
| **New** |	|	| **3.009 ms (-1%)** | **0.0307 ms** | **312.35 KB (0%)** |
| Old |Run|dromaeo-object-array|32.452 ms|0.4877 ms|96211.86 KB|
| **New** |	|	| **32.070 ms (-1%)** | **0.2978 ms** | **96211.69 KB (0%)** |
| Old |Run|droma(...)egexp [21]|93.176 ms|1.8621 ms|152871.14 KB|
| **New** |	|	| **92.373 ms (-1%)** | **1.8454 ms** | **152179.04 KB (0%)** |
| Old |Run|droma(...)tring [21]|241.801 ms|8.5015 ms|1315758.23 KB|
| **New** |	|	| **237.291 ms (-2%)** | **9.9897 ms** | **1315752.98 KB (0%)** |
| Old |Run|droma(...)ase64 [21]|35.643 ms|0.5987 ms|2268.38 KB|
| **New** |	|	| **33.913 ms (-5%)** | **0.6555 ms** | **2268.24 KB (0%)** |


## Jint.Benchmark.SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Run|3d-cube|125.04 ms|1.457 ms|45164.77 KB|
| **New** |	|	| **122.41 ms (-2%)** | **1.492 ms** | **45166.97 KB (0%)** |
| Old |Run|3d-morph|109.38 ms|1.025 ms|46258.08 KB|
| **New** |	|	| **110.58 ms (+1%)** | **0.773 ms** | **46258.12 KB (0%)** |
| Old |Run|3d-raytrace|118.38 ms|1.454 ms|86418.77 KB|
| **New** |	|	| **115.00 ms (-3%)** | **0.895 ms** | **86419.7 KB (0%)** |
| Old |Run|access-binary-trees|65.11 ms|0.838 ms|62249.23 KB|
| **New** |	|	| **64.55 ms (-1%)** | **0.982 ms** | **62249.15 KB (0%)** |
| Old |Run|access-fannkuch|271.38 ms|2.336 ms|103.84 KB|
| **New** |	|	| **281.50 ms (+4%)** | **5.519 ms** | **103.82 KB (0%)** |
| Old |Run|access-nbody|126.09 ms|0.829 ms|53295.3 KB|
| **New** |	|	| **132.79 ms (+5%)** | **1.889 ms** | **53295.29 KB (0%)** |
| Old |Run|access-nsieve|89.84 ms|0.730 ms|17142.67 KB|
| **New** |	|	| **92.39 ms (+3%)** | **1.201 ms** | **17142.64 KB (0%)** |
| Old |Run|bitop(...)-byte [24]|96.59 ms|1.071 ms|61948.44 KB|
| **New** |	|	| **100.41 ms (+4%)** | **1.883 ms** | **61948.44 KB (0%)** |
| Old |Run|bitops-bits-in-byte|147.83 ms|2.501 ms|40544.59 KB|
| **New** |	|	| **148.46 ms (0%)** | **1.495 ms** | **40544.67 KB (0%)** |
| Old |Run|bitops-bitwise-and|83.29 ms|0.607 ms|55938.95 KB|
| **New** |	|	| **82.73 ms (-1%)** | **0.935 ms** | **55939.01 KB (0%)** |
| Old |Run|bitops-nsieve-bits|135.38 ms|1.805 ms|53916.36 KB|
| **New** |	|	| **137.89 ms (+2%)** | **1.576 ms** | **53916.37 KB (0%)** |
| Old |Run|contr(...)rsive [21]|75.75 ms|0.938 ms|92772.97 KB|
| **New** |	|	| **76.70 ms (+1%)** | **1.207 ms** | **92772.97 KB (0%)** |
| Old |Run|crypto-aes|80.93 ms|0.835 ms|10749.68 KB|
| **New** |	|	| **82.92 ms (+2%)** | **1.135 ms** | **10731.43 KB (0%)** |
| Old |Run|crypto-md5|70.33 ms|0.915 ms|82176.43 KB|
| **New** |	|	| **71.67 ms (+2%)** | **0.974 ms** | **82176.42 KB (0%)** |
| Old |Run|crypto-sha1|68.12 ms|0.613 ms|68861.22 KB|
| **New** |	|	| **70.07 ms (+3%)** | **0.335 ms** | **68861.17 KB (0%)** |
| Old |Run|date-format-tofte|66.77 ms|1.147 ms|54569.06 KB|
| **New** |	|	| **67.45 ms (+1%)** | **0.720 ms** | **54161.32 KB (-1%)** |
| Old |Run|date-format-xparb|38.01 ms|0.755 ms|26400.13 KB|
| **New** |	|	| **38.31 ms (+1%)** | **0.758 ms** | **26400.21 KB (0%)** |
| Old |Run|math-cordic|217.84 ms|3.686 ms|86859.77 KB|
| **New** |	|	| **216.07 ms (-1%)** | **3.070 ms** | **86859.41 KB (0%)** |
| Old |Run|math-partial-sums|77.78 ms|0.975 ms|49367.85 KB|
| **New** |	|	| **77.27 ms (-1%)** | **0.703 ms** | **49367.9 KB (0%)** |
| Old |Run|math-spectral-norm|80.72 ms|1.310 ms|56617.52 KB|
| **New** |	|	| **81.59 ms (+1%)** | **1.179 ms** | **56617.52 KB (0%)** |
| Old |Run|regexp-dna|100.39 ms|1.011 ms|16824.85 KB|
| **New** |	|	| **101.04 ms (+1%)** | **1.529 ms** | **16829.54 KB (0%)** |
| Old |Run|string-base64|52.29 ms|0.432 ms|3156.24 KB|
| **New** |	|	| **51.43 ms (-2%)** | **0.657 ms** | **3156.21 KB (0%)** |
| Old |Run|string-fasta|118.04 ms|2.052 ms|104786.09 KB|
| **New** |	|	| **111.50 ms (-6%)** | **1.019 ms** | **104785.97 KB (0%)** |
| Old |Run|string-tagcloud|53.91 ms|0.788 ms|42042.57 KB|
| **New** |	|	| **53.61 ms (-1%)** | **0.734 ms** | **42218.87 KB (0%)** |
| Old |Run|string-unpack-code|52.66 ms|0.787 ms|73810.69 KB|
| **New** |	|	| **50.69 ms (-4%)** | **0.662 ms** | **73810.13 KB (0%)** |
| Old |Run|strin(...)input [21]|43.06 ms|0.581 ms|20687.73 KB|
| **New** |	|	| **42.64 ms (-1%)** | **0.365 ms** | **20687.13 KB (0%)** |